Fate point reroll for learning and ptgs
Refactor roll.js to break out into multiple smaller files.
Fate points now correctly distribute when using beginner's luck.
Retroactive test and ptgs success counting when a fate reroll turns
a failure into a success.
